Listen on localhost only (Linux)
With localhost in the base address, the service was listening on all interfaces
(0.0.0.0). With 127.0.0.1, it only listen on the loopback interface.
Tested with mono 5.10.1.20

Note: For some reason the service now seems to listen in IPv6 only
(seen in netstat -plnet). The client CLI manages to connect anyway.

Reference: https://bugzilla.xamarin.com/show_bug.cgi?id=275